# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend service for the Cryptocash Battery on the TON network.",
        "title": "Battery implementation for TON network",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/firebase/status": {
            "get": {
                "description": "Verifies the Firestore connection with a live round-trip and reports whether Firebase is reachable",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "firebase"
                ],
                "summary": "Firebase connection status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.FirebaseStatusResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/handler.FirebaseStatusResponse"
                        }
                    }
                }
            }
        },
...
        "/print/execute": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Reserves charges, sends the buffer to the new wallet (buffer-first), settles or releases on result",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "print"
                ],
                "summary": "Execute a printed wallet",
                "parameters": [
                    {
                        "description": "quote id",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.printExecuteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.PrintExecuteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/print/quote": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Computes the buffer and charge cost for printing a new wallet and stores the quote",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "print"
                ],
                "summary": "Quote a printed (gift) wallet",
                "parameters": [
                    {
                        "description": "destination address",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.printQuoteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.PrintQuoteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/print/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the status of a print job",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "print"
                ],
                "summary": "Print job status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "print job id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.PrintStatusResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/relayer/status": {
            "get": {
                "description": "Shows the relayer wallet address and its on-chain balance and seqno (reads TON testnet)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "relayer"
                ],
                "summary": "Relayer status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.RelayerStatusResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/ton-proof/dev-sign": {
            "get": {
                "description": "TEST ONLY — signs a ton_proof for a fixed test wallet from a payload, so /ton-proof/check can be exercised without a real wallet. Enabled only when DEV_SIGN=true. Response carries header X-Dev-Route: TO-BE-DELETED.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dev"
                ],
                "summary": "[TO BE DELETED] Dev-only ton_proof signer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "payload from /ton-proof/payload",
                        "name": "payload",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "app domain (default localhost)",
                        "name": "domain",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tonproof.Request"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handler.FirebaseStatusResponse": {
            "type": "object",
            "properties": {
                "connected": {
                    "type": "boolean"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handler.PrintExecuteResponse": {
            "type": "object",
            "properties": {
                "charge": {
                    "type": "integer"
                },
                "quoteId": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "txHash": {
                    "type": "string"
                }
            }
        },
        "handler.PrintQuoteResponse": {
            "type": "object",
            "properties": {
                "bufferTON": {
                    "type": "string"
                },
                "charge": {
                    "type": "integer"
                },
                "expiresAt": {
                    "type": "integer"
                },
                "quoteId": {
                    "type": "string"
                }
            }
        },
        "handler.PrintStatusResponse": {
            "type": "object",
            "properties": {
                "bufferTON": {
                    "type": "string"
                },
                "charge": {
                    "type": "integer"
                },
                "quoteId": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "toAddress": {
                    "type": "string"
                },
                "txHash": {
                    "type": "string"
                }
            }
        },
...
        "handler.RelayerStatusResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "balanceNano": {
                    "type": "integer"
                },
                "deployed": {
                    "type": "boolean"
                },
                "error": {
                    "type": "string"
                },
                "online": {
                    "type": "boolean"
                },
                "seqno": {
                    "type": "integer"
                }
            }
        },
...
        "handler.printExecuteRequest": {
            "type": "object",
            "properties": {
                "quoteId": {
                    "type": "string"
                }
            }
        },
        "handler.printQuoteRequest": {
            "type": "object",
            "properties": {
                "toAddress": {
                    "type": "string"
                }
            }
        },
...
        "tonproof.Domain": {
            "type": "object",
            "properties": {
                "length_bytes": {
                    "type": "integer"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "tonproof.Proof": {
            "type": "object",
            "properties": {
                "domain": {
                    "$ref": "#/definitions/tonproof.Domain"
                },
                "payload": {
                    "type": "string"
                },
                "signature": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
        "tonproof.Request": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "network": {
                    "type": "string"
                },
                "proof": {
                    "$ref": "#/definitions/tonproof.Proof"
                },
                "public_key": {
                    "type": "string"
                },
                "wallet_state_init": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "AdminToken": {
            "type": "apiKey",
            "name": "X-Admin-Token",
            "in": "header"
        },
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}